Former-commit-id: 77dad4e68abf6c7e5c4fb2f077a261fcfccd0dc1
Former-commit-id: d6c895453677a7af3bf18b4e4b1f8bc9ce0ab7e9

+++ backend/src/main/java/com/ssafy/i5i/hotelAPI/domain/hotel/entity/Attraction.java
@@ -54,10 +54,10 @@ public class Attraction {
     private Integer gugunCode;
 
     @Column(name = "latitude")
-    private BigDecimal latitude;
+    private Double latitude;
 
     @Column(name = "longitude")
-    private BigDecimal longitude;
+    private Double longitude;
 
     @Column(name = "mlevel")
     private String mlevel;
